+++ src/components/physics_body.rs
@@ -1,7 +1,7 @@
 use rapier3d::prelude::*;
 
 use crate::math::Vec3;
-use crate::physics_world::PhysicsWorld;
+use crate::physics::PhysicsWorld;
 
 pub struct PhysicsBody {
     handle: RigidBodyHandle,
